@@ -40,90 +40,96 @@
     "version": "conventional-changelog -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md",
     "lint": "ng lint",
     "package-mac": "electron-packager . --overwrite --platform=darwin --arch=x64 --icon=assets/icons/mac/icon.icns --prune=true --out=release-builds",
-    "package-win": "electron-packager . electron-tutorial-app --overwrite --asar --platform=win32 --arch=x64 --icon=assets/icons/win/icon.ico --prune=true --out=release-builds --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName=\"Electron Tutorial App\"",    
+    "package-win": "electron-packager . electron-tutorial-app --overwrite --asar --platform=win32 --arch=x64 --prune=true --out=release-builds --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName=\"Electron Tutorial App\"",
     "package-linux": "electron-packager . electron-tutorial-app --overwrite --asar --platform=linux --arch=x64 --icon=assets/icons/png/1024x1024.png --prune=true --out=release-builds"
- 
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "9.2.0",
-    "@angular-devkit/build-angular": "0.1000.5",
-    "@angular-eslint/builder": "0.0.1-alpha.32",
-    "@angular-eslint/eslint-plugin": "0.0.1-alpha.32",
-    "@angular-eslint/eslint-plugin-template": "0.0.1-alpha.32",
-    "@angular-eslint/template-parser": "0.0.1-alpha.32",
-    "@angular/cli": "10.0.5",
-    "@angular/common": "10.0.9",
-    "@angular/compiler": "10.0.9",
-    "@angular/compiler-cli": "10.0.9",
-    "@angular/core": "10.0.9",
-    "@angular/forms": "10.0.9",
-    "@angular/language-service": "10.0.9",
-    "@angular/platform-browser": "10.0.9",
-    "@angular/platform-browser-dynamic": "10.0.9",
-    "@angular/router": "10.0.9",
+    "@angular-builders/custom-webpack": "10.0.0",
+    "@angular-devkit/architect": "^0.1001",
+    "@angular-devkit/build-angular": "0.1001.0",
+    "@angular-eslint/builder": "0.3.0-beta.1",
+    "@angular-eslint/eslint-plugin": "0.3.0-beta.1",
+    "@angular-eslint/eslint-plugin-template": "0.3.0-beta.1",
+    "@angular-eslint/template-parser": "0.3.0-beta.1",
+    "@angular/cli": "10.1.0",
+    "@angular/common": "10.1.0",
+    "@angular/compiler": "10.1.0",
+    "@angular/compiler-cli": "10.1.0",
+    "@angular/core": "10.1.0",
+    "@angular/forms": "10.1.0",
+    "@angular/language-service": "10.1.0",
+    "@angular/platform-browser": "10.1.0",
+    "@angular/platform-browser-dynamic": "10.1.0",
+    "@angular/router": "10.1.0",
     "@ngx-translate/core": "13.0.0",
     "@ngx-translate/http-loader": "6.0.0",
-    "@types/jasmine": "3.5.12",
+    "@types/jasmine": "3.5.14",
     "@types/jasminewd2": "2.0.8",
-    "@types/mocha": "8.0.2",
-    "@types/node": "12.12.6",
-    "@typescript-eslint/eslint-plugin": "3.9.0",
-    "@typescript-eslint/eslint-plugin-tslint": "3.9.0",
-    "@typescript-eslint/parser": "3.9.0",
+    "@types/mocha": "8.0.3",
+    "@types/node": "14.6.4",
+    "@typescript-eslint/eslint-plugin": "4.0.1",
+    "@typescript-eslint/eslint-plugin-tslint": "4.0.1",
+    "@typescript-eslint/parser": "4.0.1",
     "chai": "4.2.0",
-    "conventional-changelog-cli": "2.0.34",
+    "conventional-changelog-cli": "2.1.0",
     "core-js": "3.6.5",
     "cross-env": "7.0.2",
-    "electron": "9.2.0",
+    "electron": "10.1.1",
     "electron-builder": "22.8.0",
     "electron-reload": "1.5.0",
-    "eslint": "7.6.0",
+    "electron-winstaller": "4.0.1",
+    "electron-wix-msi": "3.0.0",
+    "eslint": "7.8.1",
     "eslint-plugin-import": "2.22.0",
+    "firebase-tools": "^8.10.0",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^7.3.3",
+    "inquirer-autocomplete-prompt": "^1.1.0",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "5.0.2",
-    "karma": "5.1.1",
+    "karma": "5.2.1",
     "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-electron": "6.3.1",
-    "karma-jasmine": "3.3.1",
+    "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.5.4",
-    "mocha": "8.1.1",
+    "mocha": "8.1.3",
     "npm-run-all": "4.1.5",
+    "open": "^7.2.1",
     "rxjs": "6.6.2",
     "spectron": "11.1.0",
-    "ts-node": "8.10.2",
+    "ts-node": "9.0.0",
     "tslib": "2.0.1",
-    "typescript": "3.9.7",
-    "wait-on": "5.0.1",
+    "typescript": "4.0.2",
+    "wait-on": "5.2.0",
     "webdriver-manager": "12.1.7",
-    "zone.js": "0.10.3",
-    "@angular-devkit/architect": ">= 0.900 < 0.1100",
-    "firebase-tools": "^8.0.0",
-    "fuzzy": "^0.1.3",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.0.1",
-    "open": "^7.0.3"
+    "zone.js": "0.11.1"
   },
   "engines": {
     "node": ">=10.13.0"
   },
   "dependencies": {
-    "@angular/animations": "10.0.9",
-    "@angular/cdk": "10.1.3",
+    "@angular/animations": "10.1.0",
+    "@angular/cdk": "10.2.0",
     "@angular/fire": "6.0.2",
-    "@swimlane/ngx-charts": "14.0.0",
+    "@angular/forms": "^10.0.0 || ^11.0.0-0",
+    "@swimlane/ngx-charts": "16.0.0",
     "angular-split": "4.0.0",
     "bootstrap": "4.5.2",
-    "firebase": "7.18.0",
+    "directory-tree": "2.2.4",
+    "firebase": "7.19.1",
     "iv-viewer": "2.0.1",
     "jquery": "3.5.1",
+    "ng-image-slider": "2.6.4",
     "ng2-dragula": "2.1.1",
     "ng2-image-viewer": "3.0.5",
     "ngx-filter-pipe": "2.1.2",
     "ngx-infinite-scroll": "9.0.0",
     "ngx-pagination": "5.0.0",
     "ngx-spinner": "10.0.1",
-    "popper.js": "1.14.3",
+    "ngx-treeview": "10.0.2",
+    "popper.js": "1.16.1",
     "rxjs-compat": "6.6.2",
-    "sweetalert2": "9.17.1"
+    "sweetalert2": "9.17.2",
+    "walkdir": "0.4.1"
   }
 }
