@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import { Component, OnDestroy, OnInit} from '@angular/core';
+import { isNil, remove, reverse } from 'lodash';
 import { ElectronService } from './core/services';
 import { TranslateService } from '@ngx-translate/core';
 import { AuthService } from './auth/auth.service';
@@ -8,7 +9,8 @@ import { AuthService } from './auth/auth.service';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
+  
   constructor(
     private electronService: ElectronService,
     private translate: TranslateService,
@@ -16,4 +18,7 @@ export class AppComponent {
   ) {
     this.translate.setDefaultLang('en');
   }
+  ngOnInit(): void {
+  }
 }
+
